update
Message,Topicの投稿ウインドウを開いたままメインウインドウが操作できるようにしてみる

+++ Lair/Windows/ConnectionsSettingsWindow.xaml.cs
@@ -1319,11 +1319,11 @@ namespace Lair.Windows
                 int count = ConnectionsSettingsWindow.GetStringToInt(_bandwidthConnectionCountTextBox.Text);
                 _lairManager.ConnectionCountLimit = Math.Max(Math.Min(count, 100), 12);
 
-                long bandwidthLimit = (long)NetworkConverter.FromSizeString("0");
+                int bandwidthLimit = (int)NetworkConverter.FromSizeString("0");
 
                 try
                 {
-                    bandwidthLimit = (long)NetworkConverter.FromSizeString(_bandwidthLimitTextBox.Text);
+                    bandwidthLimit = (int)NetworkConverter.FromSizeString(_bandwidthLimitTextBox.Text);
                 }
                 catch (Exception)
                 {
